Fix live dashboard not updating - persist phase state for late-connecting clients
## Problem
Live dashboard showed 'Initializing...' and 'Waiting for connection...' even after
analysis progressed through multiple phases. Events sent before browser connected
to SSE stream were lost.

## Root Cause
- Phase events were sent to event_queue before browser finished loading
- When browser connected to SSE (/api/events), earlier events were already consumed
- No mechanism to replay current state for late-connecting clients

## Solution
1. Added global current_phase_state dict to persist latest phase info
2. Updated send_phase() to update current_phase_state when sending events
3. Modified SSE /api/events endpoint to immediately send current phase
   state when any client connects

## Files Modified
- mira/visualization/live_server.py
  - Added current_phase_state global (line 36-42)
  - Updated send_phase() to persist state (line 597-602)
  - Modified events() to send phase on connect (line 101-108)

## Result
Now when browser connects (even mid-analysis), it immediately receives
the current phase, updating the display from 'Initializing...' to
the actual current phase (e.g., 'PHASE 4/7: SUBSPACE ANALYSIS')

+++ mira/core/model_wrapper.py
@@ -105,7 +105,7 @@ class ModelWrapper:
             self.model = AutoModelForCausalLM.from_pretrained(
                 load_path,
                 cache_dir=cache_dir,
-                torch_dtype=self.dtype,
+                dtype=self.dtype,  # Changed from torch_dtype to dtype (deprecated warning fix)
                 trust_remote_code=True,
                 attn_implementation="eager",  # Enable output_attentions
             ).to(self.device)
